Creating a branch
When I want to change something on github, 

CLICK THE LITTLE PENCIL to edit, you NEED a new branch, but clicking the pencil  ASSUMES that you, the code you just looked at, as a branch. Time goes forward, the sapling moves towards the sky, the branches that were at the top. Time is alway running away.